[update] キャプチャ処理を分割
UmaWindowCaptureにクラスを分割
