+++ UmaUmaChecker/CaptureMode.h
@@ -1,0 +1,7 @@
+#pragma once
+
+enum class CaptureMode
+{
+	BitBlt = 0,
+	WindowGraphicsCapture,
+};
